+++ index.js
@@ -1,5 +1,5 @@
 if (process.env.NODE_ENV !== 'production') {
-    require('dotenv').config()
+  require('dotenv').config()
 }
 const express = require('express')
 const app = express()
@@ -10,6 +10,7 @@ const cors = require('cors')
 const morgan = require('morgan')
 const history = require('connect-history-api-fallback')
 
+const admin = require('./routes/admin')
 const project = require('./routes/projects.js')
 const exhibition = require('./routes/exhibitions.js')
 const inventory = require('./routes/inventory.js')
@@ -20,23 +21,23 @@ const payment = require('./routes/payment.js')
 const dbUrl = process.env.MONGO_URL
 
 mongoose.connect(dbUrl, {
-    useNewUrlParser: true,
-    useCreateIndex: true,
-    useUnifiedTopology: true,
-    useFindAndModify: false
+  useNewUrlParser: true,
+  useCreateIndex: true,
+  useUnifiedTopology: true,
+  useFindAndModify: false
 })
 
 const db = mongoose.connection
 db.on('error', console.error.bind(console, 'connection error:'))
 db.once('open', () => {
-    console.log('Database connected')
+  console.log('Database connected')
 })
 
 // ======== MIDDLEWARE ========
 // cors
 const corsOpts = {
-    origin: process.env.CORS_ORIGIN,
-    credentials: true
+  origin: process.env.CORS_ORIGIN,
+  credentials: true
 }
 app.use(cors(corsOpts))
 
@@ -50,6 +51,7 @@ app.use(history())
 app.use(morgan('dev'))
 
 // ======== ROUTES ========
+app.use('/api/admin', admin)
 app.use('/api/projects', project)
 app.use('/api/exhibitions', exhibition)
 app.use('/api/inventory', inventory)
@@ -58,5 +60,5 @@ app.use('/api/payment-intent', payment)
 const port = process.env.PORT || 5000
 
 app.listen(port, () => {
-    console.log(`'ERE ME NOW ON ${port}`)
+  console.log(`'ERE ME NOW ON ${port}`)
 })
