+++ requests.rest
@@ -1,1 +1,1 @@
-POST http://localhost:5000/api/payment HTTP/1.1
+GET http://localhost:5000/api/admin/inventory HTTP/1.1
